Fix references
