+++ crates/rattler_networking/Cargo.toml
@@ -18,12 +18,16 @@ blocking = ['reqwest/blocking']
 
 [dependencies]
 anyhow = "1.0.75"
+bisection = "0.1.0"
 dirs = "5.0.1"
 fslock = "0.2.1"
+futures = "0.3.28"
+http-content-range = "0.1.2"
 itertools = "0.11.0"
 keyring = "2.0.5"
 lazy_static = "1.4.0"
 libc = "0.2.148"
+memmap2 = "0.9.0"
 netrc-rs = "0.1.2"
 once_cell = "1.18.0"
 reqwest = { version = "0.11.22", default-features = false }
@@ -31,6 +35,9 @@ retry-policies = { version = "0.2.0", default-features = false }
 serde = "1.0.188"
 serde_json = "1.0.107"
 thiserror = "1.0.49"
+tokio = { version = "1.32.0", default-features = false, features = ["macros", "test-util"] }
+tokio-stream = { version = "0.1.14", features = ["sync"] }
+tokio-util = { version = "0.7.9", features = ["compat"] }
 tracing = "0.1.37"
 url = "2.4.1"
 
@@ -39,5 +46,10 @@ getrandom = { version = "0.2.10", features = ["js"] }
 
 [dev-dependencies]
 anyhow = "1.0.75"
+assert_matches = "1.5.0"
+async_zip = { version = "0.0.15", default-features = false, features = ["tokio"] }
+tower-http = { version = "0.4.4", default-features = false, features = ["fs"] }
+axum = { version = "0.6.20", default-features = false, features = ["tokio"] }
 insta = { version = "1.33.0", features = ["json"] }
+rstest = { version = "0.18.2" }
 tempfile = "3.8.0"
